Add functionality for repeated report names
